baselines/ChordGNN/requirements.txt
pytorch_lightning==1.5.10
torch_scatter==2.1.1
partitura==1.3.1
pandas==1.2.3
rotograd==0.1.5
tqdm==4.62.3
music21==8.1.0
GitPython==3.1.27
wandb==0.12.2
matplotlib>=3.4.0
